[REFAC] example screens (padding, routing)
