+++ example/lib/core/routes/app_router.dart
@@ -1,8 +1,8 @@
-import 'package:example/screens/box_example_screen.dart';
-import 'package:example/screens/circle_example_screen.dart';
-import 'package:example/screens/highlight_example_screen.dart';
+import 'package:example/screens/examples/box_example_screen.dart';
+import 'package:example/screens/examples/circle_example_screen.dart';
+import 'package:example/screens/examples/highlight_example_screen.dart';
 import 'package:example/screens/home_screen.dart';
-import 'package:example/screens/underline_example_screen.dart';
+import 'package:example/screens/examples/underline_example_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:go_router/go_router.dart';
